collector/requirements.txt
scapy
pika
ujson
python-dotenv